+++ dl4j-cv-labs/src/main/java/ai/certifai/solution/segmentation/CustomLabelGenerator.java
@@ -15,7 +15,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package ai.certifai.training.segmentation.cell;
+package ai.certifai.solution.segmentation;
 
 import org.datavec.api.io.labels.PathLabelGenerator;
 import org.datavec.api.writable.NDArrayWritable;
@@ -28,14 +28,17 @@ import org.slf4j.LoggerFactory;
 import java.io.File;
 import java.io.IOException;
 import java.net.URI;
+import java.util.List;
 
-public class CustomLabelGenerator implements PathLabelGenerator {
+public class CustomLabelGenerator implements PathLabelGenerator{
     private static Logger log = LoggerFactory.getLogger(CustomLabelGenerator.class);
 
     private final int height;
     private final int width;
     private final int channels;
     private final NativeImageLoader imageLoader;
+    private static List<org.nd4j.linalg.primitives.Pair<String, String>> replacement;
+
 
     //DIRECTORY STRUCTURE:
     //Here is the directory structure
@@ -54,29 +57,40 @@ public class CustomLabelGenerator implements PathLabelGenerator {
         return false;
     }
 
-    public CustomLabelGenerator(int height, int width, int channels) {
+    public CustomLabelGenerator(int height, int width, int channels, List replacement ){
         this.height = height;
         this.width = width;
         this.channels = channels;
         this.imageLoader = new NativeImageLoader(this.height, this.width, this.channels);
+        this.replacement = replacement;
     }
 
     // This custom label generator finds labels for each input image by replacing one the folders (inputs >> masks) in the path string.
     @Override
     public Writable getLabelForPath(String path) {
-        try {
-            String labelPath = path.replace("inputs", "masks");
 
-            NDArrayWritable label = new NDArrayWritable(imageLoader.asMatrix(new File(labelPath)));
+        String labelPath = path;
+        try
+        {
+
+            for (int i=0; i<replacement.size(); i++) {
+                labelPath = labelPath.replace(replacement.get(i).getKey(), replacement.get(i).getValue());
+            }
+
+//            System.out.println(labelPath);
+
+            NDArrayWritable label = new NDArrayWritable(imageLoader.asMatrix(new File(labelPath)) );
 
             INDArray labelINDArray = label.get();
 
             // normalise to 0-1 scale
-            label.set(labelINDArray.div(255));
+            label.set( labelINDArray.div(255));
 
-            return label;
+            return label ;
 
-        } catch (IOException ioe) {
+        }
+        catch (IOException ioe)
+        {
             ioe.printStackTrace();
             return null;
         }
@@ -87,5 +101,5 @@ public class CustomLabelGenerator implements PathLabelGenerator {
         return this.getLabelForPath(uri.getPath());
     }
 
-
 }
+
