+++ dl4j-cv-labs/src/main/java/ai/certifai/solution/segmentation/cell/CellDataSetIterator.java
@@ -17,7 +17,7 @@
 
 package ai.certifai.solution.segmentation.cell;
 
-import ai.certifai.solution.segmentation.imageUtils.CustomLabelGenerator;
+import ai.certifai.solution.segmentation.CustomLabelGenerator;
 import ai.certifai.Helper;
 import ai.certifai.utilities.DataUtilities;
 import org.datavec.api.io.filters.BalancedPathFilter;
