+++ dl4j-cv-labs/src/main/java/ai/certifai/training/segmentation/CustomLabelGenerator.java
@@ -15,7 +15,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package ai.certifai.solution.segmentation.imageUtils;
+package ai.certifai.training.segmentation;
 
 import org.datavec.api.io.labels.PathLabelGenerator;
 import org.datavec.api.writable.NDArrayWritable;
