+++ dl4j-cv-labs/src/main/java/ai/certifai/training/segmentation/cell/CellDataSetIterator.java
@@ -17,24 +17,31 @@
 
 package ai.certifai.training.segmentation.cell;
 
+import ai.certifai.training.segmentation.CustomLabelGenerator;
 import ai.certifai.Helper;
 import ai.certifai.utilities.DataUtilities;
-import org.datavec.image.transform.ImageTransform;
-import org.deeplearning4j.datasets.datavec.RecordReaderDataSetIterator;
 import org.datavec.api.io.filters.BalancedPathFilter;
 import org.datavec.api.split.FileSplit;
 import org.datavec.api.split.InputSplit;
 import org.datavec.image.loader.NativeImageLoader;
 import org.datavec.image.recordreader.ImageRecordReader;
+import org.datavec.image.transform.ImageTransform;
+import org.deeplearning4j.datasets.datavec.RecordReaderDataSetIterator;
 import org.nd4j.linalg.dataset.api.preprocessor.DataNormalization;
 import org.nd4j.linalg.dataset.api.preprocessor.ImagePreProcessingScaler;
+import org.nd4j.linalg.primitives.Pair;
+import org.slf4j.Logger;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Paths;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Random;
 
 public class CellDataSetIterator {
+    private static final Logger log = org.slf4j.LoggerFactory.getLogger(CellDataSetIterator.class);
+
     private static final int height = 224;
     private static final int width = 224;
     private static final int channels = 1;
@@ -42,7 +49,10 @@ public class CellDataSetIterator {
     private static final Random random = new Random(seed);
     private static String inputDir;
     private static String downloadLink;
-    private static CustomLabelGenerator labelMaker = new CustomLabelGenerator(height, width, channels);
+    private static List<Pair<String, String>> replacement = Arrays.asList(
+            new org.nd4j.linalg.primitives.Pair<>("inputs", "masks")
+    );
+    private static CustomLabelGenerator labelMaker = new CustomLabelGenerator(height, width, channels, replacement);
     private static InputSplit trainData, valData;
     private static int batchSize;
 
@@ -50,17 +60,14 @@ public class CellDataSetIterator {
     private static DataNormalization scaler = new ImagePreProcessingScaler(0, 1);
     private static ImageTransform transform;
 
-    public CellDataSetIterator() throws IOException {
+    public CellDataSetIterator() {
     }
 
     //This method instantiates an ImageRecordReader and subsequently a RecordReaderDataSetIterator based on it
-    private static RecordReaderDataSetIterator makeIterator(InputSplit split, boolean training) throws IOException {
+    private static RecordReaderDataSetIterator makeIterator(InputSplit split) throws IOException {
         ImageRecordReader recordReader = new ImageRecordReader(height, width, channels, labelMaker);
-        if (training && transform != null) {
-            recordReader.initialize(split, transform);
-        } else {
-            recordReader.initialize(split);
-        }
+        //        Both train and val iterator need the preprocessing of converting RGB to Grayscale
+        recordReader.initialize(split, transform);
         RecordReaderDataSetIterator iter = new RecordReaderDataSetIterator(recordReader, batchSize, 1, 1, true);
         iter.setPreProcessor(scaler);
 
@@ -68,11 +75,11 @@ public class CellDataSetIterator {
     }
 
     public static RecordReaderDataSetIterator trainIterator() throws IOException {
-        return makeIterator(trainData, true);
+        return makeIterator(trainData);
     }
 
     public static RecordReaderDataSetIterator valIterator() throws IOException {
-        return makeIterator(valData, false);
+        return makeIterator(valData);
     }
 
     public static void setup(int batchSizeArg, double trainPerc, ImageTransform imageTransform) throws IOException {
@@ -93,10 +100,22 @@ public class CellDataSetIterator {
 
         File dataZip = new File(Paths.get(inputDir, "data-science-bowl-2018", "data-science-bowl-2018.zip").toString());
         File classFolder = new File(Paths.get(inputDir, "data-science-bowl-2018", "data-science-bowl-2018").toString());
+
         if (!dataZip.exists()) {
-            if (DataUtilities.downloadFile(downloadLink, dataZip.getAbsolutePath())) {
-                DataUtilities.extractZip(dataZip.getAbsolutePath(), classFolder.getAbsolutePath());
+            log.info(String.format("Downloading %s from %s.", dataZip.getAbsolutePath(), downloadLink));
+            DataUtilities.downloadFile(downloadLink, dataZip.getAbsolutePath());
+        }
+
+        if (!classFolder.exists()) {
+
+            if (!Helper.getCheckSum(dataZip.getAbsolutePath())
+                    .equalsIgnoreCase(Helper.getPropValues("dataset.segmentationCell.hash"))) {
+                System.out.println("Downloaded file is incomplete");
+                System.exit(0);
             }
+
+            log.info(String.format("Extracting %s into %s.", dataZip.getAbsolutePath(), classFolder.getAbsolutePath()));
+            DataUtilities.extractZip(dataZip.getAbsolutePath(), classFolder.getAbsolutePath());
         }
 
         batchSize = batchSizeArg;
